Added Flytrap, Prefab & Screen Fade (Missing Art)
The flytrap script, along with a basic prefab has been added to the project, the script is contained within the trigger in the prefab, and requires a reference to the game manager within the scene, as well as the fade panel in the GUI, the player, and the output location.

Will be updated when art assets are available.